Create tagcloud.js
This JavaScript file  is used to make names of activities present in the style of the tag-cloud.

If you want use this file you need pay attention to div which's id is container1 in index.html. Besides, you should add jQuery to your website, as follows:
<script type="text/javascript" src="./script/jquery-2.1.1.min.js"></script>

you can see this effect in activity page(1) of www.batbat8.com.